+++ packages/lib/universal-sql/src/client-duckdb/browser.js
@@ -55,6 +55,8 @@ export async function initDB() {
 
 		// use an intermediate variable to prevent db from being a not-ready database
 		const _db = new AsyncDuckDB(logger, worker);
+		window[Symbol.for('EVIDENCE_QUERY_ENGINE')] = _db;
+
 		await _db.instantiate(DUCKDB_CONFIG.mainModule);
 		db = _db;
 
@@ -67,6 +69,12 @@ export async function initDB() {
 			}
 		});
 		connection = await db.connect();
+
+		// https://duckdb.org/2024/09/09/announcing-duckdb-110.html#breaking-sql-changes
+		await connection.query('SET ieee_floating_point_ops = false;');
+		// https://duckdb.org/2024/02/13/announcing-duckdb-0100.html#breaking-sql-changes
+		await connection.query('SET old_implicit_casting = true;');
+
 		resolveInit();
 	} catch (e) {
 		rejectInit(e);