+++ packages/ui/core-components/src/lib/atoms/shadcn/badge/index.js
@@ -2,12 +2,12 @@ import { tv } from 'tailwind-variants';
 
 export { default as Badge } from './badge.svelte';
 export const badgeVariants = tv({
-	base: 'inline-flex items-center rounded-md border px-2.5 py-0.5 text-xs font-semibold transition-colors focus:outline-none focus:ring-2 focus:ring-gray-400 focus:ring-offset-2 select-none',
+	base: 'inline-flex items-center rounded-md border border-base-300 px-2.5 py-0.5 text-xs font-semibold transition-colors focus:outline-none focus:ring-2 focus:ring-base-content-muted focus:ring-offset-2 select-none',
 	variants: {
 		variant: {
-			default: 'border-transparent bg-gray-900 text-white shadow hover:bg-gray-900/80',
-			secondary: 'border-transparent bg-gray-100 text-gray-900 hover:bg-gray-100/80',
-			destructive: 'border-transparent bg-red-500 text-red-50 shadow hover:bg-red-500/80',
+			default: 'border-transparent bg-base-300 text-base-content shadow hover:bg-base-300/80',
+			destructive:
+				'border-transparent bg-negative text-negative-content shadow hover:bg-negative/80',
 			outline: 'text-foreground'
 		}
 	},