+++ packages/ui/core-components/src/lib/unsorted/ui/Modal.svelte
@@ -19,34 +19,36 @@
 
 <button
 	on:click={() => isOpen()}
-	class="font-ui font-normal text-base no-underline border my-2 rounded-lg py-2 px-3 inline-block transition hover:ease-in hover:border-blue-200 hover:bg-blue-100 hover:no-underline hover:text-gray-800 focus:text-gray-900 focus:border-blue-200"
+	class="font-ui font-normal text-base no-underline border border-info/20 bg-info/10 my-5 rounded-lg py-2 px-3 inline-block transition hover:ease-in hover:border-info/50 hover:bg-info/20 hover:no-underline focus:border-info"
 >
 	{buttonText}
 </button>
 {#if open}
 	<div class="modal z-50 fixed w-full h-full top-0 left-0 flex items-center justify-center lg:p-0">
-		<div class="fixed inset-0 bg-gray-900 opacity-50" />
-		<div class="bg-white w-4/5 max-w-lg mx-auto rounded-lg shadow-xl z-50 overflow-y-auto font-ui">
+		<div class="fixed inset-0 bg-base-100/70" />
+		<div
+			class="bg-base-100 w-4/5 max-w-lg mx-auto rounded-lg shadow-xl z-50 overflow-y-auto font-ui"
+		>
 			<div
 				class="relative flex justify-between items-center {title.trim() != ''
-					? ' bg-gray-100 py-4 px-10 text-xl font-bold break-normal text-center'
+					? ' bg-base-200 py-4 px-10 text-xl font-bold break-normal text-center'
 					: ''}"
 			>
 				{#if title.trim() != ''}
 					{title}
 					<button
-						class="flex items-center justify-center border-none hover:bg-gray-200"
+						class="flex items-center justify-center border-none hover:bg-base-300"
 						on:click={() => isOpen()}
 					>
-						<Icon src={X} class="text-gray-600 w-6 h-6" />
+						<Icon src={X} class="w-6 h-6" />
 					</button>
 				{/if}
 				{#if title.trim() == ''}
 					<button
-						class=" flex items-center justify-center border-none p-0.5 absolute right-0 top-0 ml-4"
+						class="flex items-center justify-center border-none p-0.5 absolute right-0 top-0 ml-4"
 						on:click={() => isOpen()}
 					>
-						<Icon src={X} class="text-gray-600 w-6 h-6 mt-2 mr-2 hover:bg-gray-200" />
+						<Icon src={X} class="w-6 h-6 mt-2 mr-2 hover:bg-base-300" />
 					</button>
 				{/if}
 			</div>