+++ packages/ui/core-components/src/lib/unsorted/viz/map/BaseMap.svelte
@@ -59,8 +59,6 @@
 
 	let internalError = evidenceMap.internalError;
 
-	$: console.log($internalError);
-
 	$: if ($internalError !== undefined) {
 		error = $internalError;
 	}
@@ -93,7 +91,7 @@
 </script>
 
 {#if error}
-	<ErrorChart {error} {chartType} />
+	<ErrorChart {error} title={chartType} />
 {:else}
 	<div class="relative break-inside-avoid">
 		{#if title}
@@ -114,7 +112,7 @@
 
 		{#if !$allGeoJsonLoaded}
 			<div
-				class="absolute left-0 right-0 bottom-0 w-full *:m-0 bg-white"
+				class="absolute left-0 right-0 bottom-0 w-full *:m-0 bg-base-100"
 				style="height: {height}px"
 			>
 				<Skeleton />
@@ -133,7 +131,7 @@
 	}
 
 	div :global(.leaflet-popup-content-wrapper) {
-		background-color: white;
+		background-color: var(--base-100);
 		box-shadow: 0 1px 3px rgb(0, 0, 0, 0.4);
 		border-radius: 4px;
 		font-family: 'Inter', sans-serif;
@@ -145,7 +143,7 @@
 	}
 
 	div :global(.leaflet-popup-tip) {
-		background-color: white;
+		background-color: var(--base-100);
 		display: none;
 	}
 
@@ -161,7 +159,7 @@
 	}
 
 	div :global(.leaflet-tooltip::before) {
-		background-color: white;
+		background-color: var(--base-100);
 		display: none;
 	}
 </style>