+++ packages/ui/core-components/src/lib/unsorted/viz/scatter/ScatterPlot.svelte
@@ -3,8 +3,12 @@
 </script>
 
 <script>
+	import { getThemeStores } from '../../../themes/themes.js';
+
 	import Chart from '../core/Chart.svelte';
 	import Scatter from './Scatter.svelte';
+
+	const { resolveColor, resolveColorsObject, resolveColorPalette } = getThemeStores();
 
 	export let data = undefined;
 	export let x = undefined;
@@ -36,9 +40,15 @@
 	export let yMax = undefined;
 
 	export let shape = undefined;
+
 	export let fillColor = undefined;
+	$: fillColorStore = resolveColor(fillColor);
+
 	export let opacity = undefined; // opacity of both fill and outline (ECharts limitation)
+
 	export let outlineColor = undefined;
+	$: outlineColorStore = resolveColor(outlineColor);
+
 	export let outlineWidth = undefined;
 	export let pointSize = undefined;
 	export let chartAreaHeight = undefined;
@@ -51,7 +61,9 @@
 
 	let useTooltip = true;
 
-	export let colorPalette = undefined;
+	export let colorPalette = 'default';
+	$: colorPaletteStore = resolveColorPalette(colorPalette);
+
 	export let echartsOptions = undefined;
 	export let seriesOptions = undefined;
 	export let printEchartsConfig = false;
@@ -62,7 +74,10 @@
 	export let renderer = undefined;
 	export let downloadableData = undefined;
 	export let downloadableImage = undefined;
+
 	export let seriesColors = undefined;
+	$: seriesColorsStore = resolveColorsObject(seriesColors);
+
 	export let seriesOrder = undefined;
 
 	export let connectGroup = undefined;
@@ -101,7 +116,7 @@
 	{chartType}
 	{sort}
 	{chartAreaHeight}
-	{colorPalette}
+	colorPalette={colorPaletteStore}
 	{echartsOptions}
 	{seriesOptions}
 	{printEchartsConfig}
@@ -111,13 +126,13 @@
 	{downloadableData}
 	{downloadableImage}
 	{connectGroup}
-	{seriesColors}
+	seriesColors={seriesColorsStore}
 >
 	<Scatter
 		{shape}
-		{fillColor}
+		fillColor={fillColorStore}
 		{opacity}
-		{outlineColor}
+		outlineColor={outlineColorStore}
 		{outlineWidth}
 		{pointSize}
 		{useTooltip}