+++ packages/ui/core-components/src/lib/unsorted/viz/table/TableRow.svelte
@@ -8,11 +8,11 @@
 	} from '@evidence-dev/component-utilities/formatting';
 	import TableCell from './TableCell.svelte';
 	import chroma from 'chroma-js';
-	import { uiColours } from '@evidence-dev/component-utilities/colours';
 	import { Icon } from '@steeze-ui/svelte-icon';
 	import { ChevronRight } from '@steeze-ui/tabler-icons';
 	import Sparkline from '../core/_Sparkline.svelte';
 	import { addBasePath } from '@evidence-dev/sdk/utils/svelte';
+	import { getThemeStores } from '../../../themes/themes.js';
 
 	export let displayedData = undefined;
 	export let rowShading = undefined;
@@ -52,6 +52,8 @@
 
 		await goto(addBasePath(row[link]));
 	};
+
+	const { theme } = getThemeStores();
 </script>
 
 {#each displayedData as row, i}
@@ -86,31 +88,31 @@
 			{@const color_domain =
 				column.colorBreakpoints ??
 				(column.colorMid ? [column_min, column.colorMid, column_max] : [column_min, column_max])}
-			{@const color_scale = column.colorPalette
-				? chroma.scale(column.colorPalette).domain(color_domain).nodata('white')
+			{@const color_scale = column.colorScale
+				? chroma.scale(column.colorScale).domain(color_domain).nodata($theme.colors['base-100'])
 				: ''}
 			{@const cell_color =
-				column.contentType === 'colorscale' && is_nonzero && column.colorPalette
+				column.contentType === 'colorscale' && is_nonzero && column.colorScale
 					? column.scaleColumn
 						? color_scale(row[column.scaleColumn]).hex()
 						: color_scale(row[column.id]).hex()
 					: ''}
 			{@const font_color = column.redNegatives
 				? row[column.id] < 0
-					? 'rgb(220 38 38)'
+					? $theme.colors.negative
 					: ''
-				: column.contentType === 'colorscale' && is_nonzero && column.colorPalette
-					? chroma.contrast(cell_color, uiColours.grey999) <
-						chroma.contrast(cell_color, 'white') + 0.5
-						? 'white'
-						: uiColours.grey999
+				: column.contentType === 'colorscale' && is_nonzero && column.colorScale
+					? chroma.contrast(cell_color, $theme.colors['base-content']) <
+						chroma.contrast(cell_color, $theme.colors['base-100']) + 0.5
+						? $theme.colors['base-100']
+						: $theme.colors['base-content']
 					: ''}
 			{@const bottom_border =
 				i !== displayedData.length - 1 &&
 				rowLines &&
 				column.contentType === 'colorscale' &&
 				is_nonzero &&
-				column.colorPalette
+				column.colorScale
 					? `1px solid ${chroma(cell_color).darken(0.5)}`
 					: ''}
 			<TableCell
@@ -148,7 +150,7 @@
 						<a
 							href={addBasePath(row[column.id])}
 							target={column.openInNewTab ? '_blank' : ''}
-							class="text-blue-600 hover:text-blue-700 transition-colors duration-200"
+							class="text-primary hover:brightness-110 transition-colors duration-200"
 						>
 							{#if column.linkLabel != undefined}
 								<!-- if the linklabel is a column name, display that column -->
@@ -316,13 +318,13 @@
 	</tr>
 {/each}
 
-<style>
+<style lang="postcss">
 	.row-lines {
-		border-bottom: thin solid var(--grey-200);
+		@apply border-b border-base-300;
 	}
 
 	.shaded-row {
-		background-color: var(--grey-50);
+		@apply bg-base-200;
 	}
 
 	*:focus {
@@ -334,7 +336,6 @@
 	}
 
 	.row-link:hover {
-		--tw-bg-opacity: 1;
-		background-color: rgb(239 246 255 / var(--tw-bg-opacity));
+		@apply bg-base-200;
 	}
 </style>